+++ packages/http/src/xhr2.d.ts
@@ -1,0 +1,4 @@
+
+declare module 'xhr2' {
+    export const XMLHttpRequest: typeof globalThis.XMLHttpRequest;
+}